+++ NEWS.md
@@ -7,6 +7,11 @@
 - New function `derive_var_relative_flag()` for flagging observations before or
 after a condition is fulfilled (#1453)
 
+- The new function `create_period_dataset()` for creating a reference dataset
+for subperiods, periods, or phases from the ADSL dataset was added. The
+reference dataset can be used to create subperiod, period, and phase variables
+in OCCDS and BDS datasets. (#1477)
+
 - A `print()` method was added for all S3 objects defined by admiral, e.g.,
 `date_source()`, `dthcaus_source()`, ... (#858)